Overload vim-addons directory.
This simple change allows to specify vim-addons directory in .vimrc
instead of hardcoding it in the function itself. This is especially
useful if the user wants to tar-ball whole $HOME/.vim directory as a
single config without including all of the third-party repositories
that will checkout from repositories in new environment any way.